tensorrt/yolov3_onnx/requirements.txt
numpy>=1.15.1
onnx==1.2.3
pycuda>=2017.1.1
Pillow>=5.2.0
wget>=3.2